Day 9
Use Long in place of Int for memory to support
larger values.

Wrap the memory array to support out of bounds
reads and writes.

Support relative mode (note that the parameter to
instruction 9 is a delta).

Split out IntcodeComputer.

+++ src/main/scala/intcode/IntcodeProgram.scala
@@ -3,11 +3,11 @@ package intcode
 
 object IntcodeProgram {
 
-  def fromString(input: String): Vector[Int] = {
-    input.split(",").map(_.toInt).toVector
+  def fromString(input: String): Vector[Long] = {
+    input.split(",").map(_.toLong).toVector
   }
 
-  def fromLines(input: Vector[String]): Vector[Int] = {
+  def fromLines(input: Vector[String]): Vector[Long] = {
     IntcodeProgram.fromString(input.mkString(""))
   }
